Delete test utils that have moved to over_react_test
+ And move all remaining utils out of the “wsd_test_utils” directory
+ Consumers of this lib have no idea what wsd is / means.
