+++ test/over_react/component/abstract_transition_test.dart
@@ -18,10 +18,10 @@ import 'dart:async';
 import 'dart:html';
 
 import 'package:over_react/over_react.dart';
+import 'package:over_react_test/over_react_test.dart';
 import 'package:test/test.dart';
 
 import '../../test_util/test_util.dart';
-import '../../wsd_test_util/validation_util_helpers.dart';
 
 /// Main entry point for AbstractTransition testing
 main() {
