+++ test/over_react/component/resize_sensor_test.dart
@@ -20,13 +20,9 @@ import 'dart:html';
 
 import 'package:platform_detect/platform_detect.dart';
 import 'package:over_react/over_react.dart';
+import 'package:over_react_test/over_react_test.dart';
 import 'package:react/react_dom.dart' as react_dom;
 import 'package:test/test.dart';
-
-import '../../test_util/test_util.dart';
-import '../../wsd_test_util/common_component_tests.dart';
-import '../../wsd_test_util/validation_util_helpers.dart';
-import '../../wsd_test_util/zone.dart';
 
 void main() {
   group('ResizeSensor', () {
