+++ test/over_react/util/test_mode_test.dart
@@ -18,7 +18,7 @@ import 'package:over_react/src/component_declaration/component_base.dart' as com
 import 'package:over_react/src/util/test_mode.dart';
 import 'package:test/test.dart';
 
-/// Main entry point for enableTestMode and disabledTestMode testing
+/// Main entry point for [enableTestMode] and [disableTestMode] testing
 main() {
   test('enableTestMode and disableTestMode set UiProps.testMode as expected', () {
     disableTestMode();
